Add `SoloWanderLocationSystem` and `PackWanderLocationSystem` for modular wander behavior
- Extracted `SoloWanderLocationSystem` and `PackWanderLocationSystem` into dedicated files for individual and pack-based wander logic.
- Simplified `UpdateWanderLocation` by removing embedded `SoloWanderLocationSystem` and `PackWanderLocationSystem` code.
- Updated `UniversalRenderPipelineGlobalSettings` with additional debugger resources.
