+++ Assets/Systems/IAUS/Components/Components/States/Movement States/Wander/SoloWanderLocationSystem.cs
@@ -1,0 +1,54 @@
+﻿using System.Collections.Generic;
+using Components.MovementSystem;
+using DreamersIncStudio.GAIACollective;
+using IAUS.ECS.Systems;
+using Unity.Entities;
+using Unity.Mathematics;
+using Unity.Transforms;
+using Random = UnityEngine.Random;
+
+namespace IAUS.ECS.Component
+{
+    [UpdateInGroup(typeof(IAUSUpdateGroup))]
+    public partial class SoloWanderLocationSystem : WanderLocationSystemBase
+    {
+        protected override void OnUpdate()
+        {
+            Entities
+                .WithStructuralChanges()
+                .WithoutBurst()
+                .WithNone<PackMember>()
+                .ForEach((Entity entity,
+                    ref LocalTransform transform,
+                    ref WanderQuadrant wander,
+                    ref UpdateWanderLocationTag tag,
+                    ref Movement move) =>
+                {
+                    ProcessEntityTemplate(entity, ref transform, ref wander, ref tag, ref move);
+                })
+                .Run();
+        }
+
+        protected override float3 ComputeTravelPosition(Entity entity,
+            ref LocalTransform transform,
+            ref WanderQuadrant wander,
+            ref Movement move)
+        {
+            if (wander.WanderNeighborQuadrants)
+            {
+                // Choose among neighbor quadrants (outline keeps original selection behavior)
+                var positions = new List<float3>
+                {
+                    GetWanderPoint(transform.Position, wander.HashKey + 1),
+                    GetWanderPoint(transform.Position, wander.HashKey - 1),
+                    GetWanderPoint(transform.Position, wander.HashKey)
+                };
+
+
+                return positions[Random.Range(0, 2)];
+            }
+
+            return GetWanderPoint(transform.Position, wander.HashKey);
+        }
+    }
+}
