+++ Assets/Systems/IAUS/Components/Components/States/Movement States/Wander/UpdateWanderLocation.cs
@@ -1,15 +1,12 @@
 using Components.MovementSystem;
 using DreamersInc.QuadrantSystems;
 using IAUS.ECS.Systems;
-using System.Collections.Generic;
-using DreamersIncStudio.GAIACollective;
 using Unity.Entities;
 using Unity.Mathematics;
 using Unity.Transforms;
 using UnityEngine;
 using UnityEngine.AI;
 using Utilities;
-using Random = UnityEngine.Random;
 
 
 namespace IAUS.ECS.Component
@@ -79,82 +76,4 @@ namespace IAUS.ECS.Component
                    path.status == NavMeshPathStatus.PathComplete;
         }
     }
-    [UpdateInGroup(typeof(IAUSUpdateGroup))]
-    public partial class SoloWanderLocationSystem : WanderLocationSystemBase
-    {
-        protected override void OnUpdate()
-        {
-            Entities
-                .WithStructuralChanges()
-                .WithoutBurst()
-                .WithNone<PackMember>()
-                .ForEach((Entity entity,
-                    ref LocalTransform transform,
-                    ref WanderQuadrant wander,
-                    ref UpdateWanderLocationTag tag,
-                    ref Movement move) =>
-                {
-                    ProcessEntityTemplate(entity, ref transform, ref wander, ref tag, ref move);
-                })
-                .Run();
-        }
-
-        protected override float3 ComputeTravelPosition(Entity entity,
-            ref LocalTransform transform,
-            ref WanderQuadrant wander,
-            ref Movement move)
-        {
-            if (wander.WanderNeighborQuadrants)
-            {
-                // Choose among neighbor quadrants (outline keeps original selection behavior)
-                var positions = new List<float3>
-                {
-                    GetWanderPoint(transform.Position, wander.HashKey + 1),
-                    GetWanderPoint(transform.Position, wander.HashKey - 1),
-                    GetWanderPoint(transform.Position, wander.HashKey)
-                };
-
-             
-                return positions[Random.Range(0, 2)];
-            }
-
-            return GetWanderPoint(transform.Position, wander.HashKey);
-        }
-        [UpdateInGroup(typeof(SimulationSystemGroup))]
-        public partial class PackWanderLocationSystem : WanderLocationSystemBase
-        {
-            protected override void OnUpdate()
-            {
-                Entities
-                    .WithStructuralChanges()
-                    .WithoutBurst()
-                    .WithAll<PackMember>()
-                    .ForEach((Entity entity,
-                        ref LocalTransform transform,
-                        ref WanderQuadrant wander,
-                        ref UpdateWanderLocationTag tag,
-                        ref Movement move,
-                        in PackMember packMember) =>
-                    {
-                        // Maintain herd center update behavior
-                        wander.WanderCenterPoint = EntityManager.GetComponentData<Pack>(packMember.PackEntity).HerdCenter;
-
-                        ProcessEntityTemplate(entity, ref transform, ref wander, ref tag, ref move);
-                    })
-                    .Run();
-            }
-
-            protected override float3 ComputeTravelPosition(Entity entity,
-                ref LocalTransform transform,
-                ref WanderQuadrant wander,
-                ref Movement move)
-            {
-                // Minimal outline: reuse current hash-based wander point.
-                // You could refine this to bias toward wander.WanderCenterPoint if desired.
-                return GetWanderPoint(transform.Position, wander.HashKey);
-            }
-        }
-
-    }
-
 }
