+++ Assets/UniversalRenderPipelineGlobalSettings.asset
@@ -135,6 +135,7 @@ MonoBehaviour:
       - rid: 6165830400758775901
       - rid: 6165830400758775902
       - rid: 6165830400758775903
+      - rid: 6165830400758775904
     m_RuntimeSettings:
       m_List: []
   m_AssetVersion: 8
@@ -740,6 +741,11 @@ MonoBehaviour:
       data:
         m_version: 0
     - rid: 6165830400758775903
+      type: {class: RenderingDebuggerRuntimeResources, ns: UnityEngine.Rendering,
+        asm: Unity.RenderPipelines.Core.Runtime}
+      data:
+        m_version: 0
+    - rid: 6165830400758775904
       type: {class: RenderingDebuggerRuntimeResources, ns: UnityEngine.Rendering,
         asm: Unity.RenderPipelines.Core.Runtime}
       data:
